Move GitHub FS back to it's package
