+++ internal/evaluator/evaluator.go
@@ -10,7 +10,7 @@ import (
 	"github.com/cirruslabs/cirrus-cli/pkg/larker"
 	"github.com/cirruslabs/cirrus-cli/pkg/larker/fs"
 	"github.com/cirruslabs/cirrus-cli/pkg/larker/fs/failing"
-	"github.com/cirruslabs/cirrus-cli/pkg/larker/fs/git"
+	"github.com/cirruslabs/cirrus-cli/pkg/larker/fs/github"
 	"github.com/cirruslabs/cirrus-cli/pkg/larker/fs/memory"
 	"github.com/cirruslabs/cirrus-cli/pkg/parser"
 	"github.com/cirruslabs/cirrus-cli/pkg/parser/parsererror"
@@ -101,7 +101,7 @@ func fsFromEnvironment(env map[string]string) (fs fs.FileSystem, err error) {
 		return
 	}
 
-	return git.NewGitHub(owner, repo, reference, token)
+	return github.New(owner, repo, reference, token)
 }
 
 func (r *ConfigurationEvaluatorServiceServer) EvaluateConfig(
