+++ pkg/larker/loader/module_fs.go
@@ -5,6 +5,7 @@ import (
 	"errors"
 	"github.com/cirruslabs/cirrus-cli/pkg/larker/fs"
 	"github.com/cirruslabs/cirrus-cli/pkg/larker/fs/git"
+	"github.com/cirruslabs/cirrus-cli/pkg/larker/fs/github"
 	"regexp"
 )
 
@@ -111,7 +112,7 @@ func findLocatorFS(
 	case gitHubLocation:
 		token, _ := env["CIRRUS_REPO_CLONE_TOKEN"]
 
-		ghFS, err := git.NewGitHub(l.Owner, l.Name, l.Revision, token)
+		ghFS, err := github.New(l.Owner, l.Name, l.Revision, token)
 		if err != nil {
 			return nil, "", err
 		}
